+++ mypyc/irbuild/builder.py
@@ -64,6 +64,7 @@ from mypyc.irbuild.nonlocalcontrol import (
 from mypyc.irbuild.context import FuncInfo, ImplicitClass
 from mypyc.irbuild.mapper import Mapper
 from mypyc.irbuild.ll_builder import LowLevelIRBuilder
+from mypyc.irbuild.util import is_constant
 
 GenFunc = Callable[[], None]
 
@@ -1187,3 +1188,36 @@ class IRBuilder:
 
     def error(self, msg: str, line: int) -> None:
         self.errors.error(msg, self.module_path, line)
+
+
+def gen_arg_defaults(builder: IRBuilder) -> None:
+    """Generate blocks for arguments that have default values.
+
+    If the passed value is an error value, then assign the default
+    value to the argument.
+    """
+    fitem = builder.fn_info.fitem
+    for arg in fitem.arguments:
+        if arg.initializer:
+            target = builder.environment.lookup(arg.variable)
+
+            def get_default() -> Value:
+                assert arg.initializer is not None
+
+                # If it is constant, don't bother storing it
+                if is_constant(arg.initializer):
+                    return builder.accept(arg.initializer)
+
+                # Because gen_arg_defaults runs before calculate_arg_defaults, we
+                # add the static/attribute to final_names/the class here.
+                elif not builder.fn_info.is_nested:
+                    name = fitem.fullname + '.' + arg.variable.name
+                    builder.final_names.append((name, target.type))
+                    return builder.add(LoadStatic(target.type, name, builder.module_name))
+                else:
+                    name = arg.variable.name
+                    builder.fn_info.callable_class.ir.attributes[name] = target.type
+                    return builder.add(
+                        GetAttr(builder.fn_info.callable_class.self_reg, name, arg.line))
+            assert isinstance(target, AssignmentTargetRegister)
+            builder.assign_if_null(target, get_default, arg.initializer.line)
